+++ ZeShmouttsAssets/Blizzard API for Unity/Scripts/WoW Game Data API/Mythic Keystone Dungeon API/Endpoint_WoW_GameData_MythicKeystoneDungeon.cs
@@ -1,0 +1,60 @@
+﻿using System;
+using System.Collections;
+using ZeShmouttsAssets.BlizzardAPI.JSON;
+
+namespace ZeShmouttsAssets.BlizzardAPI
+{
+	/// <summary>
+	/// Interface for working with the Blizzard API inside Unity.
+	/// </summary>
+	public static partial class BlizzardAPI
+	{
+		/// <summary>
+		/// API endpoints related to World of Warcraft game data (items, spells, etc.).
+		/// Reference : https://develop.battle.net/documentation/world-of-warcraft/game-data-apis
+		/// </summary>
+		public static partial class WowGameData
+		{
+			/// <summary>
+			/// Coroutine that retrieves a WoW Mythic Keystone dungeon.
+			/// </summary>
+			/// <param name="dungeonId">The ID of the dungeon.</param>
+			/// <param name="action_Result">Action to execute with the data once retrieved and converted.</param>
+			/// <param name="ifModifiedSince">Adds a request header to check if the document has been modified since this date (in HTML format), which will return an empty response body if it's older.</param>
+			/// <param name="action_LastModified">Action to execute with the date of the last server-side modification to the document.</param>
+			/// <param name="region">The region of the data to retrieve.</param>
+			/// <returns></returns>
+			public static IEnumerator GetMythicKeystoneDungeon(int dungeonId, Action<WowMythicKeystoneDungeon_JSON> action_Result, string ifModifiedSince = null, Action<string> action_LastModified = null, BattleNetRegion region = BattleNetRegion.UnitedStates)
+			{
+				string path = string.Format("/data/wow/mythic-keystone/dungeon/{0}", dungeonId);
+				yield return SendRequest(region, namespaceDynamic, path, action_Result, ifModifiedSince, action_LastModified);
+			}
+		}
+	}
+}
+
+namespace ZeShmouttsAssets.BlizzardAPI.JSON
+{
+	/// <summary>
+	/// JSON structure for World of Warcraft Mythic Keystone dungeons.
+	/// </summary>
+	[Serializable]
+	public class WowMythicKeystoneDungeon_JSON : Object_Json
+	{
+		public LinkStruct _links;
+
+		public int id;
+		public LocalizedString name;
+		public NameIdStruct map;
+		public SlugStruct zone;
+		public RefNameIdStruct dungeon;
+
+		[Serializable]
+		public struct KeystoneUpgrade
+		{
+			public int upgrade_level;
+			public int qualifying_duration;
+		}
+		public KeystoneUpgrade[] keystone_upgrades;
+	}
+}