+++ ZeShmouttsAssets/Blizzard API for Unity/Scripts/WoW Game Data API/Mythic Keystone Dungeon API/Endpoint_WoW_GameData_MythicKeystoneSeason.cs
@@ -1,0 +1,50 @@
+﻿using System;
+using System.Collections;
+using ZeShmouttsAssets.BlizzardAPI.JSON;
+
+namespace ZeShmouttsAssets.BlizzardAPI
+{
+	/// <summary>
+	/// Interface for working with the Blizzard API inside Unity.
+	/// </summary>
+	public static partial class BlizzardAPI
+	{
+		/// <summary>
+		/// API endpoints related to World of Warcraft game data (items, spells, etc.).
+		/// Reference : https://develop.battle.net/documentation/world-of-warcraft/game-data-apis
+		/// </summary>
+		public static partial class WowGameData
+		{
+			/// <summary>
+			/// Coroutine that retrieves a WoW Mythic Keystone season.
+			/// </summary>
+			/// <param name="seasonId">The ID of the Mythic Keystone season.</param>
+			/// <param name="action_Result">Action to execute with the data once retrieved and converted.</param>
+			/// <param name="ifModifiedSince">Adds a request header to check if the document has been modified since this date (in HTML format), which will return an empty response body if it's older.</param>
+			/// <param name="action_LastModified">Action to execute with the date of the last server-side modification to the document.</param>
+			/// <param name="region">The region of the data to retrieve.</param>
+			/// <returns></returns>
+			public static IEnumerator GetMythicKeystoneSeason(int seasonId, Action<WowMythicKeystoneSeason_JSON> action_Result, string ifModifiedSince = null, Action<string> action_LastModified = null, BattleNetRegion region = BattleNetRegion.UnitedStates)
+			{
+				string path = string.Format("/data/wow/mythic-keystone/season/{0}", seasonId);
+				yield return SendRequest(region, namespaceDynamic, path, action_Result, ifModifiedSince, action_LastModified);
+			}
+		}
+	}
+}
+
+namespace ZeShmouttsAssets.BlizzardAPI.JSON
+{
+	/// <summary>
+	/// JSON structure for World of Warcraft Mythic Keystone seasons.
+	/// </summary>
+	[Serializable]
+	public class WowMythicKeystoneSeason_JSON : Object_Json
+	{
+		public LinkStruct _links;
+		public int id;
+		public long start_timestamp;
+		public long end_timestamp;
+		public RefIdStruct[] periods;
+	}
+}