+++ ZeShmouttsAssets/Blizzard API for Unity/Scripts/_Common/BlizzardAPI_JSON.cs
@@ -194,5 +194,11 @@
 		public int A;
 	}
 
+	[System.Serializable]
+	public struct SlugStruct
+	{
+		public string slug;
+	}
+
 	#endregion
 }